EnsembleModule moved to be nested class of AbstractEnsemble
