+++ src/main/java/weka_uea/classifiers/ensembles/AbstractEnsemble.java
@@ -27,13 +27,11 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
-import java.util.logging.Level;
-import java.util.logging.Logger;
+import timeseriesweka.classifiers.Checkpointable;
 import timeseriesweka.classifiers.MultiThreadable;
 import timeseriesweka.classifiers.SaveParameterInfo;
+import timeseriesweka.classifiers.TestTimeContractable;
 import timeseriesweka.classifiers.TrainAccuracyEstimator;
 import timeseriesweka.classifiers.TrainTimeContractable;
 import utilities.DebugPrinting;
@@ -193,6 +191,99 @@ public abstract class AbstractEnsemble extends AbstractClassifier implements Sav
     
     
     
+    
+    
+    
+    /**
+     * Simple data type to hold a classifier and it's related information and results.
+     */
+    public static class EnsembleModule implements DebugPrinting {
+        private Classifier classifier;
+
+        private String moduleName;
+        private String parameters;
+
+        public ClassifierResults trainResults;
+        public ClassifierResults testResults;
+
+        //by default (and i imagine in the vast majority of cases) all prior weights are equal (i.e 1)
+        //however may be circumstances where certain classifiers are themselves part of 
+        //a subensemble or something
+        public double priorWeight = 1.0; 
+
+        //each module makes a vote, with a weight defined for this classifier when predicting this class 
+        //many weighting schemes will have weights for each class set to a single classifier equal, but some 
+        //will have e.g certain members being experts at classifying certain classes etc
+        public double[] posteriorWeights;
+
+        public EnsembleModule() {
+            this.moduleName = "ensembleModule";
+            this.classifier = null;
+
+            trainResults = null;
+            testResults = null;
+        }
+
+        public EnsembleModule(String moduleName, Classifier classifier, String parameters) {
+            this.classifier = classifier;
+            this.moduleName = moduleName;
+            this.parameters = parameters;
+
+            trainResults = null;
+            testResults = null;
+        }
+
+        public boolean isTrainAccuracyEstimator() {
+            return classifier instanceof TrainAccuracyEstimator;
+        }
+        
+        public boolean isTrainTimeContractable() {
+            return classifier instanceof TrainTimeContractable;
+        }
+        
+        public boolean isTestTimeContractable() {
+            return classifier instanceof TestTimeContractable;
+        }
+        
+        public boolean isMultiThreadable() {
+            return classifier instanceof MultiThreadable;
+        }
+        
+        public boolean isCheckpointable() {
+            return classifier instanceof Checkpointable;
+        }
+        
+        public String getModuleName() {
+            return moduleName;
+        }
+
+        public void setModuleName(String moduleName) {
+            this.moduleName = moduleName;
+        }
+
+        public String getParameters() {
+            return parameters;
+        }
+
+        public void setParameters(String parameters) {
+            this.parameters = parameters;
+        }
+
+        public Classifier getClassifier() {
+            return classifier;
+        }
+
+        public void setClassifier(Classifier classifier) {
+            this.classifier = classifier;
+        }
+
+        @Override
+        public String toString() {
+            return moduleName;
+        }
+    }
+
+
