+++ src/main/java/weka_uea/classifiers/ensembles/voting/MajorityConfidence.java
@@ -14,7 +14,7 @@
  */
 package weka_uea.classifiers.ensembles.voting;
 
-import weka_uea.classifiers.ensembles.EnsembleModule;
+import weka_uea.classifiers.ensembles.AbstractEnsemble.EnsembleModule;
 import weka.core.Instance;
 
 /**
