+++ src/main/java/weka_uea/classifiers/ensembles/voting/MajorityVoteByPoweredConfidence.java
@@ -14,8 +14,7 @@
  */
 package weka_uea.classifiers.ensembles.voting;
 
-import java.util.Arrays;
-import weka_uea.classifiers.ensembles.EnsembleModule;
+import weka_uea.classifiers.ensembles.AbstractEnsemble.EnsembleModule;
 import static utilities.GenericTools.indexOfMax;
 import weka.core.Instance;
