+++ src/main/java/weka_uea/classifiers/ensembles/voting/NP_MAX.java
@@ -14,7 +14,7 @@
  */
 package weka_uea.classifiers.ensembles.voting;
 
-import weka_uea.classifiers.ensembles.EnsembleModule;
+import weka_uea.classifiers.ensembles.AbstractEnsemble.EnsembleModule;
 import weka.core.Instance;
 
 /**
