+++ src/main/java/weka_uea/classifiers/ensembles/weightings/ConfusionEntropy.java
@@ -14,8 +14,7 @@
  */
 package weka_uea.classifiers.ensembles.weightings;
 
-import java.util.Arrays;
-import weka_uea.classifiers.ensembles.EnsembleModule;
+import weka_uea.classifiers.ensembles.AbstractEnsemble.EnsembleModule;
 
 /**
  *
