+++ src/main/java/weka_uea/classifiers/ensembles/weightings/FScore.java
@@ -14,7 +14,7 @@
  */
 package weka_uea.classifiers.ensembles.weightings;
 
-import weka_uea.classifiers.ensembles.EnsembleModule;
+import weka_uea.classifiers.ensembles.AbstractEnsemble.EnsembleModule;
 
 /**
  * Non uniform weighting scheme, uses F-measure to give a weighting composed of
