+++ src/main/java/weka_uea/classifiers/ensembles/weightings/ProportionalTrainAcc.java
@@ -14,7 +14,7 @@
  */
 package weka_uea.classifiers.ensembles.weightings;
 
-import weka_uea.classifiers.ensembles.EnsembleModule;
+import weka_uea.classifiers.ensembles.AbstractEnsemble.EnsembleModule;
 
 /**
  * Each modules weighting is defined as it's accuracy proportional to the other modules,
