+++ src/main/java/weka_uea/classifiers/ensembles/weightings/TrainAcc.java
@@ -14,7 +14,7 @@
  */
 package weka_uea.classifiers.ensembles.weightings;
 
-import weka_uea.classifiers.ensembles.EnsembleModule;
+import weka_uea.classifiers.ensembles.AbstractEnsemble.EnsembleModule;
 
 /**
  * Simply uses the modules train acc as it's weighting. Extension: can raise the accuracy
