+++ src/main/java/weka_uea/classifiers/ensembles/weightings/TrainAccByClass.java
@@ -14,7 +14,7 @@
  */
 package weka_uea.classifiers.ensembles.weightings;
 
-import weka_uea.classifiers.ensembles.EnsembleModule;
+import weka_uea.classifiers.ensembles.AbstractEnsemble.EnsembleModule;
 
 /**
  * Simply calculates this classifier's accuracy on each class
